requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-environ==0.12.0
django-heroku==0.3.1
dnspython==2.8.0
email-validator==2.3.0
filelock==3.19.1
idna==3.10
postgres==4.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
py3-validate-email==1.0.5.post2
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
whitenoise==6.11.0
